// src/components/ButtonPokemonType/ButtonPokemonType.d.ts
export interface ButtonPokemonTypeProps {
  classes?: any;
  type: string;
}

export enum Types {
  Colorless = "Colorless",
  Darkness = "Darkness",
  Dragon = "Dragon",
  Fairy = "Fairy",
  Fighting = "Fighting",
  Fire = "Fire",
  Grass = "Grass",
  Lightning = "Lightning",
  Metal = "Metal",
  Psychic = "Psychic",
  Water = "Water",
}